requirements-test.txt
# Test dependencies — install alongside requirements.txt
# pip install -r requirements.txt -r requirements-test.txt

pytest>=8.0
pytest-mock>=3.12
pytest-asyncio>=0.23
pytest-cov>=4.1
httpx>=0.27          # async test client for FastAPI
anyio[trio]>=4.2     # pytest-asyncio backend
playwright>=1.43     # E2E tests (run: playwright install chromium)

# Phase 1 backend deps (not yet in requirements.txt)
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
pydantic-settings>=2.1
boto3>=1.34          # S3/MinIO client
celery>=5.3
redis>=5.0
